+++ src/MapGeneration/MapEventTile.cpp
@@ -10,10 +10,10 @@ MapEventTile::MapEventTile()
 MapEventTile::MapEventTile(bool needButton, int x, int y, bool walkableTile) : MapEvent(needButton)
 {
     //ctor
-    m_rect.left = x * TileMap::GetTileWidth();
-    m_rect.top = y * TileMap::GetTileWidth();
-    m_rect.width = TileMap::GetTileWidth();
-    m_rect.height = TileMap::GetTileWidth();
+    m_rect.left = (float)(x * TileMap::GetTileWidth());
+    m_rect.top = (float)(y * TileMap::GetTileWidth());
+    m_rect.width = (float)(TileMap::GetTileWidth());
+    m_rect.height = (float)(TileMap::GetTileWidth());
     m_walkableTile = walkableTile;
 }
